+++ web/src/atoms/Banner.stories.ts
