+++ web/src/atoms/Banner.vue
