+++ web/src/atoms/index.ts
@@ -1,1 +1,2 @@
 export * from './auth'
+export { default as Banner } from './Banner.vue'
