+++ web/src/components/Waitinglist.vue
@@ -71,7 +71,7 @@
 </template>
 <script>
 import http from '../http'
-import Banner from './shared/Banner.vue'
+import { Banner } from '../atoms'
 import Spinner from './shared/Spinner.vue'
 
 export default {
