+++ web/src/components/codebase/settings/SettingsACL.vue
@@ -44,7 +44,7 @@ import { toRef, ref, watch } from 'vue'
 
 import HorizontalDivider from '../../shared/HorizontalDivider.vue'
 import Button from '../../shared/Button.vue'
-import Banner from '../../shared/Banner.vue'
+import { Banner } from '../../../atoms'
 
 export default {
   name: 'SettingsACL',
