+++ web/src/components/differ/ReviewCommentChild.vue
@@ -81,7 +81,7 @@ import { defineComponent, PropType } from 'vue'
 import { Comment } from './event'
 import time from '../../time'
 import Avatar from '../shared/Avatar.vue'
-import Banner from '../shared/Banner.vue'
+import { Banner } from '../../atoms'
 import ReviewCommentMenu from './ReviewCommentMenu.vue'
 import Button from '../shared/Button.vue'
 import TextareaAutosize from '../shared/TextareaAutosize.vue'
