+++ web/src/components/workspace/LiveDetails.vue
@@ -197,7 +197,7 @@ import Differ from '../differ/Differ.vue'
 import http from '../../http'
 import { ExternalLinkIcon, XIcon } from '@heroicons/vue/outline'
 import TooManyFilesChanged from './TooManyFilesChanged.vue'
-import Banner from '../shared/Banner.vue'
+import { Banner } from '../../atoms'
 import Avatar from '../shared/Avatar.vue'
 import Button from '../shared/Button.vue'
 import { CombinedError, gql, useMutation, useQuery } from '@urql/vue'
