+++ web/src/pages/WorkspaceHome.vue
@@ -542,7 +542,7 @@ import { FolderAddIcon } from '@heroicons/vue/outline'
 import LiveDetails, { LIVE_DETAILS_WORKSPACE } from '../components/workspace/LiveDetails.vue'
 import { MEMBER_FRAGMENT } from '../components/shared/TextareaMentions.vue'
 import http from '../http'
-import Banner from '../components/shared/Banner.vue'
+import { Banner } from '../atoms'
 import ResolveConflict from '../components/workspace/ResolveConflict.vue'
 import Button from '../components/shared/Button.vue'
 import debounce from '../debounce'
